Make test_check.py pytest friendly
